requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
simple-salesforce==1.12.6
salesforce-bulk==2.2.0
sendgrid==6.11.0
openai==1.12.0
python-dotenv==1.0.1
gunicorn==21.2.0